Now ensures that libipmf so created with MPI FC
While this is not required with MPICH, OpenMPI uses a chain of weak
symbols to implement fortran pmpi routines. Unless the correct compiler
is used when creating the ipmf SO, these will not be resolved properly
at link time.

This requires adding a dummy EXTRA fortran libtool source dep to
src/Makefile.am, and manually patching libtool to handle idiosyncracies
of how some fortran compilers (such as gfortran) put a space between -l
and the library name when verbosely printing their link line. The latter
is done in the new "bootstrap.sh" script, which patches m4/libtool.m4
(which is generated on calls to autoreconf, such as in the script).