Rank remapping for comm != MPI_COMM_WORLD (Issue #13)
Previously, it was implicitly assumed that rank arguments to MPI
communication routines (e.g. the destination rank of an MPI_Send) were
relative to MPI_COMM_WORLD. In the case where this was not true,
incorrect ranks would be recorded (leading to incorrect adjacency
graphs, etc.). This patch now uses MPI_Group_translate_ranks in the case
where the supplied communicator is not MPI_COMM_WORLD.

Because MPI_Wait provides a rank through the MPI_Status object but does
not supply the communicator under which it is defined, the rank is no
longer recorded when intercepting calls to MPI_Wait (similar to what we
already to for MPI_Waitall, etc.).