Add a constructor function for serial applications
Add a function with the constructor attribute to initialize the IPM
library. This is needed to profile applications which do not call
MPI_Init. Previously, there was only support for serial applications
when using the IPM I/O wrappers. However, in this case, the IPM
library was only initialized when the application performed I/O.