Make IPM backwards compatible with MPI-1 and MPI-2.
This is because we added MPI-3 non-blocking collectives in commit a0119cc. We keep the code compatible by using the preprocessor in wrap_mpi_c.c to exclude MPI-3 functions when using older MPIs. I have not made the same change in wrap_mpi_f.c because non-blocking Fortran calls are not wrapped. This is not because of commit a0119cc but because of the code in https://github.com/nerscadmin/IPM/blob/a0119cc6c8cd298d2af63108193b582c5ca55346/bin/make_wrappers.in#L647-L652. I don't understand the motive for this code in make_wrappers.in yet (only that it pre-dates this github repo).