Fix uninitialized variable compiler warning.
```
report.c: In function 'compute_local_region_stats':
report.c:254:47: warning: 'hpio.bytesum' may be used uninitialized in this function [-Wmaybe-uninitialized]
       GSTATS_SET( stats->pio_GiB, hpio.bytesum/(1024.0 * 1024.0 * 1024.0), 1 );
```

This warning only appears when IPM is *not* compiled with POSIXIO monitoring. The warning is harmless because the uninitialized data is never accessed, however, it is good to remove this warning. The only time the variable is accessed is in `report_banner.c:240` when IPM is compiled with POSIXIO monitoring. The code was introduced by me in commit fad94c6.